Added Login page functionality
Added login.html page format and users.login blueprint for the login form, the get user query hasn't been tested on the database because connecting didn't seem to work so I will test and change that accordingly. Moved the blueprint registers into __main__ as it was causing an import cycle error before.