Created Quiz pages and functionality, interacting with the database
Added Quizzes page to dashboard where all the quizzes available to the users are displayed. Currently it gets all the quizzes in the database so it will need to be changed to get the ones assigned to the group the user is in. Added quiz pages to take the quizzes on with radio_fields and a results page that displays the % score they got. Decided to use flask_navigation to create the dynamic list of quizzes available to the user and had to use an app context to use it outside app.py. Quizzes require a specific format and are currently limited to 5 questions, this can easily be changed in the future but for now this is all we need.