Added key stage information for students available quizzes and database structure
KeyStage is now a database table that holds key stage information, currently it only holds the actual key stage and and it used by Group and Quiz as foreign key reference to link them when determining what quizzes are for what key stage and by extension which groups use them. The quizzes view now only shows the student quizzes that are under their groups key stage.